adjust view, implement access denied page for modify/delete action

+++ weekend-kaki/hangout/views.py
@@ -1,8 +1,8 @@
 
 from django.http.response import HttpResponseRedirect, HttpResponseRedirectBase
-from django.shortcuts import render
+from django.shortcuts import render, redirect
 from django.http import HttpResponse
-from django.contrib.auth.decorators import login_required
+from django.contrib.auth.decorators import login_required, user_passes_test
 from django.utils import timezone
 from .models import Hangout
 from .forms import HangoutViewForm, HangoutModifyForm, HangoutCreateForm
@@ -51,41 +51,55 @@ def discover(request):
     return render(request, 'hangout/discover.html', context)
 
 @login_required
-# show modify/delete options if owned hangout
 def detail(request, hangout_id):
     h = Hangout.objects.get(pk=hangout_id)
     form = HangoutViewForm(instance=h)
+
     if 'msg' in request.GET:
         msg = request.GET['msg']
     else:
         msg = None
+
+    if h.owner_id == request.user.id:
+        owned = True
+    else:
+        owned = False
+        
     context = {'hangout_id': hangout_id,
                'form': form,
                'msg': msg,
+               'owned': owned,
                }
     return render(request, 'hangout/detail.html', context)
 
 @login_required
-# test owned hangout
 def modify(request, hangout_id):
     h = Hangout.objects.get(pk=hangout_id)
-    form = HangoutModifyForm(instance=h)
-    context = {'hangout_id': hangout_id,
-                'form': form,
-                }
-    if request.method == 'GET':
-        return render(request, 'hangout/modify.html', context)
+    if h.owner_id == request.user.id:
+        form = HangoutModifyForm(instance=h)
+        context = {'hangout_id': hangout_id,
+                    'form': form,
+                    }
+        if request.method == 'GET':
+            return render(request, 'hangout/modify.html', context)
+        else:
+            form = HangoutModifyForm(request.POST, instance=h)
+            form.save()
+            return HttpResponseRedirect(reverse('hangout:detail', args=(hangout_id,))
+                                        + '?msg=success')
     else:
-        form = HangoutModifyForm(request.POST, instance=h)
-        form.save()
-        return HttpResponseRedirect(reverse('hangout:detail', args=(hangout_id,))
-                                    + '?msg=success')
+        return redirect('hangout:access-denied')
         
 
 @login_required
-# test owned hangout
 def delete(request, hangout_id):
     h = Hangout.objects.get(pk=hangout_id)
-    h.delete()
-    context = {}
-    return render(request, 'hangout/delete.html', context)
+    if h.owner_id == request.user.id:
+        h.delete()
+        context = {}
+        return render(request, 'hangout/delete.html', context)
+    else:
+        return redirect('hangout:access-denied')
+
+def access_denied(request):
+    return render(request, 'hangout/access_denied.html', {})
